Making the API more restful!
