+++ .eslintrc
@@ -22,6 +22,7 @@
     "templateStrings": true
   },
   "rules": {
+    "no-undef": 2,
     "strict": [2, "never"],
     "no-var": 2,
     "prefer-const": 2,
