+++ server.js
@@ -3,21 +3,20 @@ import http from 'http';
 import cors from 'cors';
 import RedisCache from 'services/RedisCache';
 import { green, red } from 'colors/safe';
+import router from 'resources/router';
 
 RedisCache.on('message', (message) => {
   console.log(`${green.bold('RedisCache Message:')} ${message}`);
 });
 
 RedisCache.on('error', (message) => {
-  console.log(`${red.bold('RedisCache Message:')} ${message}`);
+  console.error(`${red.bold('RedisCache Message:')} ${message}`);
 });
-
-import routes from 'routes';
 
 const server = http.createServer(
   express()
     .use(cors())
-    .use('/:region', routes)
+    .use(router)
 );
 
 server.listen(9000, ()=> {
